refactor: 断点续传检查点持久化单位由块数改为字节偏移
- CheckpointEntry.completedBlocks(int 块数) 改为 completedBytes(long 字节偏移)，与分块大小解耦，未来调整块大小不破坏旧检查点
- ReadFileCall/WriteFileCall 两端以当前块大小 floor 对齐持久化值后执行跳过（宁可重传不可漏传）
- 握手协议响应值由 int 块数改为 long 字节偏移（readLong/writeLong）
- ConfigDB 版本升至 3：v2 表 completed_blocks 按 1MB 块换算迁移为 completed_bytes
- JdkCheckpointManager JSON 字段改为 completedBytes，并兼容解析旧格式 completedBlocks
- 新增集成测试：字节偏移持久化/就地续传/非块边界对齐/ReadFileCall 跳块/旧格式兼容 全部通过

+++ HybridFileXfer-Android/app/src/main/java/top/weixiansen574/hybridfilexfer/droidcore/DroidReadFileCall.java
@@ -21,7 +21,7 @@ public class DroidReadFileCall extends ReadFileCall {
     private FileInputStream fileInputStream;
     private FileChannel channel;
 
-    public DroidReadFileCall(IIOService ioService, LinkedBlockingDeque<ByteBuffer> buffers, List<RemoteFile> files, Directory localDir, Directory remoteDir, int operateThreadCount, Map<String, Integer> checkpoints) {
+    public DroidReadFileCall(IIOService ioService, LinkedBlockingDeque<ByteBuffer> buffers, List<RemoteFile> files, Directory localDir, Directory remoteDir, int operateThreadCount, Map<String, Long> checkpoints) {
         super(buffers, files, localDir, remoteDir, operateThreadCount, checkpoints);
         this.ioService = ioService;
     }
